requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
boto3==1.34.140
botocore==1.34.140
branca==0.7.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.5.0
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
descartes==1.1.0
dnspython==2.6.1
email_validator==2.1.2
fake-useragent==1.5.1
Faker==26.0.0
fastapi==0.110.1
fastapi-login==1.10.0
fastapi-sessions==0.3.2
filelock==3.15.4
flake8==7.1.0
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
geopandas==1.0.1
google-auth==2.32.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
kafka==1.3.5
kafka-python==2.0.2
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
multidict==6.0.5
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.4
mysqlclient-pool==0.1.1
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parsel==1.9.1
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyDispatcher==2.0.7
pyflakes==3.2.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyogrio==0.9.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyproj==3.6.1
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
redis==5.0.7
requests==2.31.0
requests-file==2.1.0
rsa==4.9
s3transfer==0.10.2
schedule==1.2.2
Scrapy==2.11.2
scrapy-selenium==0.0.7
selenium==4.22.0
service-identity==24.1.0
setuptools==69.5.1
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
starlette-session==0.4.3
svgwrite==1.4.3
tldextract==5.1.2
trio==0.21.0
trio-websocket==0.11.1
Twisted==24.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
w3lib==2.2.1
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
xyzservices==2024.6.0
yarl==1.9.4
zope.interface==6.4.post2